Remove account protobuf generated Go code
All of the Go code generated from the account protobuf definitions has been removed. This includes the various Go structs for the data types defined in `account.proto` and the gRPC client and server interfaces and implementations for the account service. These types were used in the authentication and account management part of the application. This indicates a possible switch to another protocol or data format.